requirements.txt
requests
python-dotenv
openai
pygithub
